python/0170 - Web Calendar/requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
flask
flask_restful
requests
Flask-SQLAlchemy
certifi
markupsafe